requirements.txt
Flask
pandas
pytorch
matplotlib
